Move tx proof config to a separate module
