+++ ethereum_history_api/oracles/src/noir/oracles/transactionOracle/encode.ts
@@ -6,25 +6,7 @@ import { BYTE_HEX_LEN, U1_ZERO } from '../../../util/const.js';
 import { Proof } from '../../../ethereum/proof.js';
 import { padArray } from '../../../util/array.js';
 import { removeHexPrefix } from '../../../util/hex.js';
-import { getProofConfig } from '../common/proofConfig.js';
-
-// TODO: Remove this when we remove legacy fixtures
-export const LEGACY_MAX_TX_ENCODED_LEN = 525;
-export const LEGACY_MAX_TX_RLP_LEN = LEGACY_MAX_TX_ENCODED_LEN - 1;
-
-class TxProofConfig {
-  public static readonly MAX_KEY_LEN = 3;
-  public static readonly MAX_PROOF_LEVELS = 7;
-}
-
-export class TxProofConfigM extends TxProofConfig {
-  public static MAX_VALUE_LEN = 1000;
-
-  private static readonly config = getProofConfig(this.MAX_KEY_LEN, this.MAX_VALUE_LEN, this.MAX_PROOF_LEVELS);
-  public static readonly MAX_PREFIXED_KEY_NIBBLE_LEN = this.config.maxPrefixedKeyNibbleLen;
-  public static readonly MAX_LEAF_LEN = this.config.maxLeafLen;
-  public static readonly MAX_PROOF_LEN = this.config.maxProofLen;
-}
+import { LEGACY_MAX_TX_RLP_LEN, TxProofConfigM } from '../common/proofConfig/tx.js';
 
 export enum TX_OFFSETS {
   NONCE,
@@ -55,7 +37,7 @@ export function encodeTx(transaction: Transaction): ForeignCallOutput[] {
 }
 
 export function encodeTxProof(txProof: Proof): ForeignCallOutput[] {
-  const key = encodeBytes(BigInt(txProof.key), TxProofConfig.MAX_KEY_LEN);
+  const key = encodeBytes(BigInt(txProof.key), TxProofConfigM.MAX_KEY_LEN);
   const value = padArray(encodeHex(txProof.value), LEGACY_MAX_TX_RLP_LEN, ZERO_PAD_VALUE);
   const proof = encodeProof(txProof.proof, TxProofConfigM.MAX_PROOF_LEN);
   const depth = encodeField(txProof.proof.length);
