+++ ethereum_history_api/oracles/src/script/noir_fixtures/new_transaction_proof.ts
@@ -1,5 +1,5 @@
 import { Proof } from '../../ethereum/proof.js';
-import { TxProofConfigM } from '../../noir/oracles/transactionOracle/encode.js';
+import { TxProofConfigM } from '../../noir/oracles/common/proofConfig/tx.js';
 import { createNewTopLevelProofInputFixture } from './new_proof.js';
 
 export function createNewTransactionProofFixture(proof: Proof): string {
